# docs/_archive/2026-09-18/openapi.yaml
openapi: 3.1.0
info:
  title: YYC³ NovaMind API
  version: 2.0.0
  description: |
    YYC³ NovaMind 全栈 API 契约（OpenAPI 3.1）。

    本契约由 M3 MiniMax 导师于 2026-09-18 自动扫描 `app/api/**` 导出函数生成，
    覆盖 13 个路由文件 / 30 个端点，作为前后端对接的唯一权威来源。

    ## 五维契约原则
    - **时间维**：所有时间字段统一 ISO-8601 UTC
    - **空间维**：所有路径相对 baseURL（默认 `http://localhost:3218`）
    - **属性维**：所有响应含 `success: boolean`，错误码与 `lib/error-handler.ts` 对齐
    - **事件维**：所有 POST/PUT 操作记录 trace_id 用于审计
    - **关联维**：复杂资源附带 `_links` 字段指向关联资源
  contact:
    name: YYC³ Team
    email: admin@0379.email
  license:
    name: Copyright © 2025-2026 YanYuCloudCube
    identifier: YYC-Internal
servers:
  - url: http://localhost:3218
    description: 本地开发（端口 3218）
  - url: https://staging.api.yyc3.app
    description: 预发布
  - url: https://api.yyc3.app
    description: 生产
...
paths:
...
  /api/ai:
    get:
      tags: [AI]
      summary: 查询 AI 服务状态 / 提供商
      parameters:
        - in: query
          name: action
          required: true
          schema:
            type: string
            enum: [status, providers, current_provider, test_provider]
        - in: query
          name: provider
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [AI]
      summary: AI 文本生成（流式或单次）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [AI]
      summary: 切换 AI 提供商或更新配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIConfigUpdate'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ===== Image =====
  /api/analyze-image:
    post:
      tags: [Image]
      summary: 图像内容分析
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageAnalyzeRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Image]
      summary: 获取图像分析能力 / 模型元数据
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ===== Auth =====
...
components:
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    InternalError:
      description: 内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          description: 业务是否成功
        data:
          description: 业务负载（任意 JSON）
        message:
          type: string
        trace_id:
          type: string
          description: 链路追踪 ID
        timestamp:
          type: string
          format: date-time
    ApiError:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION, AUTH, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMIT, INTERNAL]
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            severity:
              type: string
              enum: [LOW, MEDIUM, HIGH, CRITICAL]
            trace_id:
              type: string
    AIRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
        provider:
          type: string
          enum: [openai, anthropic, mock]
        model:
          type: string
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 0.7
        maxTokens:
          type: integer
          minimum: 1
          maximum: 32000
        stream:
          type: boolean
          default: false
    AIConfigUpdate:
      type: object
      required: [provider]
      properties:
        provider:
          type: string
        apiKey:
          type: string
          description: 可选，覆盖环境变量
    ImageAnalyzeRequest:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
          format: uri
        prompt:
          type: string
        features:
          type: array
          items:
            type: string
            enum: [object_detection, ocr, scene_classification, color_analysis]
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: yyc3_session
      description: 会话 Cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 备用 Bearer Token
security:
  - sessionCookie: []
  - bearerAuth: []